requirements.txt
# LLM and Orchestration
langgraph>=1.1.0
langchain-core>=1.2.0
langchain-openai>=1.1.0
langchain-anthropic>=1.4.0
langchain-google-genai>=2.0.0
langgraph-checkpoint-sqlite>=3.0.0

# Execution and Containerization
docker>=7.0.0

# Memory and Storage
chromadb>=0.5.0
sentence-transformers>=2.7.0

# UI
streamlit>=1.36.0

# Utilities
python-dotenv>=1.0.0
pytest>=8.0.0
